requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
openai==1.3.0
httpx==0.24.1
pyMuPDF==1.23.8
python-docx==1.1.0
aiofiles==23.2.1
pytest==7.4.3
pytest-asyncio==0.21.1